Removed unwanted oneliner to set log level and make lots of checks using testify
