+++ examples/CallingLogExample_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/hyperjumptech/grule-rule-engine/builder"
 	"github.com/hyperjumptech/grule-rule-engine/engine"
 	"github.com/hyperjumptech/grule-rule-engine/pkg"
+	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
@@ -21,22 +22,16 @@ rule CallingLog "Calling a log" {
 )
 
 func TestCallingLog(t *testing.T) {
-	//logrus.SetLevel(logrus.DebugLevel)
 	dataContext := ast.NewDataContext()
 
 	lib := ast.NewKnowledgeLibrary()
 	ruleBuilder := builder.NewRuleBuilder(lib)
 	err := ruleBuilder.BuildRuleFromResource("CallingLog", "0.1.1", pkg.NewBytesResource([]byte(DRL)))
-	if err != nil {
-		panic(err)
-	} else {
+	assert.NoError(t, err)
 
-		knowledgeBase := lib.NewKnowledgeBaseInstance("CallingLog", "0.1.1")
+	knowledgeBase := lib.NewKnowledgeBaseInstance("CallingLog", "0.1.1")
 
-		eng1 := &engine.GruleEngine{MaxCycle: 1}
-		err := eng1.Execute(dataContext, knowledgeBase)
-		if err != nil {
-			t.Fatalf("Got error %v", err)
-		}
-	}
+	eng1 := &engine.GruleEngine{MaxCycle: 1}
+	err = eng1.Execute(dataContext, knowledgeBase)
+	assert.NoError(t, err)
 }
