+++ examples/Issue4_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/hyperjumptech/grule-rule-engine/builder"
 	"github.com/hyperjumptech/grule-rule-engine/engine"
 	"github.com/hyperjumptech/grule-rule-engine/pkg"
+	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
@@ -48,25 +49,16 @@ func TestMethodCall_Issue4(t *testing.T) {
 
 	dataContext := ast.NewDataContext()
 	err := dataContext.Add("User", user)
-	if err != nil {
-		t.Fatal(err)
-	}
+	assert.NoError(t, err)
 
 	// Prepare knowledgebase library and load it with our rule.
 	lib := ast.NewKnowledgeLibrary()
 	rb := builder.NewRuleBuilder(lib)
 	err = rb.BuildRuleFromResource("Test", "0.1.1", pkg.NewBytesResource([]byte(Rule4)))
-	if err != nil {
-		t.Fatal(err)
-	} else {
-		kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
-		eng1 := &engine.GruleEngine{MaxCycle: 3}
-		err := eng1.Execute(dataContext, kb)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if user.GetName() != "FromRuleScope4" {
-			t.Fatalf("User should be FromRuleScope4 but %s", user.GetName())
-		}
-	}
+	assert.NoError(t, err)
+	kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
+	eng1 := &engine.GruleEngine{MaxCycle: 3}
+	err = eng1.Execute(dataContext, kb)
+	assert.NoError(t, err)
+	assert.Equal(t, "FromRuleScope4", user.GetName())
 }
