+++ examples/Issue5_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/hyperjumptech/grule-rule-engine/builder"
 	"github.com/hyperjumptech/grule-rule-engine/engine"
 	"github.com/hyperjumptech/grule-rule-engine/pkg"
+	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
@@ -39,25 +40,16 @@ func TestMethodCall_Issue5(t *testing.T) {
 
 	dataContext := ast.NewDataContext()
 	err := dataContext.Add("User", user)
-	if err != nil {
-		t.Fatal(err)
-	}
+	assert.NoError(t, err)
 
 	// Prepare knowledgebase library and load it with our rule.
 	lib := ast.NewKnowledgeLibrary()
 	rb := builder.NewRuleBuilder(lib)
 	err = rb.BuildRuleFromResource("Test", "0.1.1", pkg.NewBytesResource([]byte(Rule)))
-	if err != nil {
-		t.Log(err)
-	} else {
-		eng1 := &engine.GruleEngine{MaxCycle: 5}
-		kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
-		err := eng1.Execute(dataContext, kb)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if user.GetName() != "FromRuleScope3" {
-			t.Fatalf("User should be FromRuleScope3 but %s", user.GetName())
-		}
-	}
+	assert.NoError(t, err)
+	eng1 := &engine.GruleEngine{MaxCycle: 5}
+	kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
+	err = eng1.Execute(dataContext, kb)
+	assert.NoError(t, err)
+	assert.Equal(t, "FromRuleScope3", user.GetName())
 }
