+++ examples/Issue7_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/hyperjumptech/grule-rule-engine/builder"
 	"github.com/hyperjumptech/grule-rule-engine/engine"
 	"github.com/hyperjumptech/grule-rule-engine/pkg"
+	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
@@ -41,25 +42,16 @@ func TestMethodCall_Issue7(t *testing.T) {
 
 	dataContext := ast.NewDataContext()
 	err := dataContext.Add("User", user)
-	if err != nil {
-		t.Fatal(err)
-	}
+	assert.NoError(t, err)
 
 	// Prepare knowledgebase library and load it with our rule.
 	lib := ast.NewKnowledgeLibrary()
 	rb := builder.NewRuleBuilder(lib)
 	err = rb.BuildRuleFromResource("Test", "0.1.1", pkg.NewBytesResource([]byte(Rule7)))
-	if err != nil {
-		t.Log(err)
-	} else {
-		eng1 := &engine.GruleEngine{MaxCycle: 5}
-		kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
-		err := eng1.Execute(dataContext, kb)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if user.GetName() != "FromRule" {
-			t.Fatalf("User should be FromRule but %s", user.GetName())
-		}
-	}
+	assert.NoError(t, err)
+	eng1 := &engine.GruleEngine{MaxCycle: 5}
+	kb := lib.NewKnowledgeBaseInstance("Test", "0.1.1")
+	err = eng1.Execute(dataContext, kb)
+	assert.NoError(t, err)
+	assert.Equal(t, "FromRule", user.GetName())
 }
