+++ examples/SalienceExample_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/hyperjumptech/grule-rule-engine/builder"
 	"github.com/hyperjumptech/grule-rule-engine/engine"
 	"github.com/hyperjumptech/grule-rule-engine/pkg"
+	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
@@ -104,10 +105,7 @@ func TestSalience(t *testing.T) {
 	rb := builder.NewRuleBuilder(lib)
 	byteArr := pkg.NewBytesResource([]byte(SalienceDRL))
 	err := rb.BuildRuleFromResource("Tutorial", "0.0.1", byteArr)
-
-	if err != nil {
-		panic(err)
-	}
+	assert.NoError(t, err)
 
 	engine := engine.NewGruleEngine()
 
@@ -116,19 +114,11 @@ func TestSalience(t *testing.T) {
 	for _, td := range testData {
 		dataCtx := ast.NewDataContext()
 		err := dataCtx.Add("V", td)
-		if err != nil {
-			panic(err)
-		}
+		assert.NoError(t, err)
 
 		err = engine.Execute(dataCtx, knowledgeBase)
-		if err != nil {
-			panic(err)
-		}
-
-		if td.Rating != td.Expect {
-			t.Logf("On intValue = %d, expect \"%s\" but \"%s\"", td.IntValue, td.Expect, td.Rating)
-			t.Fail()
-		}
+		assert.NoError(t, err)
+		assert.Equal(t, td.Expect, td.Rating)
 	}
 
 }
